+++ db/sqlc/models.go
@@ -11,7 +11,7 @@ import (
 type Account struct {
 	ID           int64     `json:"id"`
 	Owner        string    `json:"owner"`
-	Balance      float64   `json:"balance"`
+	Balance      int64     `json:"balance"`
 	CurrencyCode string    `json:"currency_code"`
 	CreatedAt    time.Time `json:"created_at"`
 }
@@ -19,7 +19,7 @@ type Account struct {
 type Entry struct {
 	ID        int64     `json:"id"`
 	AccountID int64     `json:"account_id"`
-	Amount    float64   `json:"amount"`
+	Amount    int64     `json:"amount"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
@@ -27,6 +27,6 @@ type Transfer struct {
 	ID            int64     `json:"id"`
 	FromAccountID int64     `json:"from_account_id"`
 	ToAccountID   int64     `json:"to_account_id"`
-	Amount        float64   `json:"amount"`
+	Amount        int64     `json:"amount"`
 	CreatedAt     time.Time `json:"created_at"`
 }
